вынес повторения в модуль, подправил шаблоны
