+++ webapp/model.py
@@ -1,3 +1,4 @@
 from flask_sqlalchemy import SQLAlchemy
  
 db = SQLAlchemy()
+
