+++ webapp/templates/deck/deck_with_cards.html
@@ -14,7 +14,7 @@
         <h5 class="card-title">{{ deck.name }}</h5>
         <p class="card-text">{{ deck.comment }}</p>
         <div class="btn-group">
-            <button type="button" class="btn btn-primary"><a class="dropdown-item" href='{{ url_for("deck.deck_study", deck_id=deck.id )}}'>Повторять карточки</a></button>
+            <button type="button" class="btn btn-primary"><a class="dropdown-item" href='{{ url_for("study.deck_study", deck_id=deck.id )}}'>Повторять карточки</a></button>
             <button type="button" class="btn btn-primary dropdown-toggle dropdown-toggle-split" data-bs-toggle="dropdown"
                 aria-expanded="false">
                 <span class="visually-hidden">Toggle Dropdown</span>
